feat: glfw window and opengl context init
To create an OpenGL context, we need an application window to draw in.
However, these operations are specific to the operating system, and
OpenGL purposefully abstracts itself from these operations to be
platform-independent. This is where GLFW comes in; it is a library that
handles window creation and input handling for us.